git-svn-id: http://subversion.ffado.org/ffado/trunk@1245 2be59082-3212-0410-8809-b0798e1608f0

+++ libffado/src/bebob/terratec/terratec_device.h
@@ -45,7 +45,7 @@ public:
     virtual ClockSource getActiveClockSource();
 
 protected:
-    virtual uint8_t getConfigurationIdSyncMode();
+    virtual uint16_t getConfigurationIdSyncMode();
 
 private:
     void updateClockSources();
